add some useful macros
